// reactAPI-SPA/src/interfaces/interfaces.tsx
export interface Switcher {
    path: string;
    destiny: string;
}


export interface Car {
    id: number;
    make: string;
    model: string;
    year: number;
    color: string;
    mileage: number;
    price: number;
    fuelType: string;
    transmission: string;
    engine: string;
    horsepower: number;
    image:string;
}